chore: move spanEndedToSpan out of span-factory
